Test for role update at username change
Here is a first draft of a function changing the role of an user when they change their username. If the new username does not follow the convention "/u/[name]", add the "New User" role. Otherwise, remove the "New User" role. I don't know if if the code has been added correctly to the bot (arguments, ...), but it should be a good starting point to implement that.